requirements.txt
# 深度学习核心依赖
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# 数据处理和可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# 生成式AI和合成数据
faker>=18.0.0

# 开发和测试
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
jupyter>=1.0.0
ipykernel>=6.23.0

# 模型评估和实验追踪
tensorboard>=2.12.0
wandb>=0.15.0

# 配置和工具
pyyaml>=6.0.0
tqdm>=4.65.0

# 时间序列分析
tslearn>=0.6.0

# 统计分析
statsmodels>=0.14.0
